Optimization of cycle search
The search for cycles has been greatly optimized. If we have already checked its dependence on cycles for some cell, then we can not check for it. I also added a check to see if the "left" and "right" fields of the operation are already written, so as not to call the function once again